Implemented alot of the first file system. (It will print alot)
I implemented the operations to do what they normally do and print when entering the operation. There are still compiler errors but I resolved the majority of them. All of the compiler errors that are left have to do with the function pointers in the fuse_operator struct. Its time for a break though!