(STILL BROKEN) Worked on make_gen a little more.
Mosto of our functionality is implemented. We have a bunch of warnings that scare me and a seg fault. We need to add debug statements and figure out how to get this workig. Once this is done we will have a working makefs that doesn't support renaming or removing files.