Almost there!! A few touch ups and bug fixes left.
Ok so everything compiles runs fine. Makefiles are generated as well. For some reason the file system crashes when trying to create the executable though, when you type make. Another thing we need to fix is in makefs_create the creat calls need to be moved. I stubbed it out. Also we need to add cases in make_generator so that deps and objs are only in the makefile if needed. Also stubbed out.